Crate pmg

This binary crate allows a user to infer the parameters of the Galaxy by optimising over its parametric model.

Modules

cli 🔒

Command-line interface

model 🔒

Model of the Galaxy

utils 🔒

Utilities

Functions

main

Run the program

serialize_n_results 🔒

Serialize the costs and errors in azimuthal velocity

write_delta_varpi 🔒

Write all mean systematical errors to a plain file

write_fit_params_to_plain 🔒

Write all fitted parameters to a plain file