Struct pmg::model::objects::Object

pub struct Object<F> {

    pub outlier: bool,
    pub from_reid: Option<bool>,
    pub name: Option<String>,
    pub obj_type: Option<String>,
    pub source: Option<String>,
    pub alpha: Option<F>,
    pub delta: Option<F>,
    pub par: Option<F>,
    pub par_e: Option<F>,
    pub par_p: Option<F>,
    pub par_m: Option<F>,
    pub v_lsr: Option<F>,
    pub v_lsr_e: Option<F>,
    pub mu_x: Option<F>,
    pub mu_x_e: Option<F>,
    pub mu_y: Option<F>,
    pub mu_y_e: Option<F>,
    pub r_h: Option<F>,
    pub r_h_ep: Option<F>,
    pub r_h_em: Option<F>,
    pub r_h_p: Option<F>,
    pub r_h_m: Option<F>,
    pub l: Option<F>,
    pub b: Option<F>,
    pub mu_l_cos_b: Option<F>,
    pub mu_b: Option<F>,
    pub r_g: Option<F>,
    pub r_g_ep: Option<F>,
    pub r_g_em: Option<F>,
    pub r_g_p: Option<F>,
    pub r_g_m: Option<F>,
    pub x: Option<F>,
    pub x_ep: Option<F>,
    pub x_em: Option<F>,
    pub x_p: Option<F>,
    pub x_m: Option<F>,
    pub y: Option<F>,
    pub y_ep: Option<F>,
    pub y_em: Option<F>,
    pub y_p: Option<F>,
    pub y_m: Option<F>,
    pub z: Option<F>,
    pub z_ep: Option<F>,
    pub z_em: Option<F>,
    pub z_p: Option<F>,
    pub z_m: Option<F>,
    pub theta: Option<F>,
    pub theta_ep: Option<F>,
    pub theta_em: Option<F>,
    pub theta_p: Option<F>,
    pub theta_m: Option<F>,
    pub theta_evel: Option<F>,
    pub theta_evel_corrected: Option<F>,
    pub v_r: Option<F>,
    pub v_r_e: Option<F>,
    pub v_l: Option<F>,
    pub v_l_ep: Option<F>,
    pub v_l_em: Option<F>,
    pub v_l_p: Option<F>,
    pub v_l_m: Option<F>,
    pub v_b: Option<F>,
    pub v_b_ep: Option<F>,
    pub v_b_em: Option<F>,
    pub v_b_p: Option<F>,
    pub v_b_m: Option<F>,
    pub u: Option<F>,
    pub u_ep: Option<F>,
    pub u_em: Option<F>,
    pub u_p: Option<F>,
    pub u_m: Option<F>,
    pub v: Option<F>,
    pub v_ep: Option<F>,
    pub v_em: Option<F>,
    pub v_p: Option<F>,
    pub v_m: Option<F>,
    pub w: Option<F>,
    pub w_ep: Option<F>,
    pub w_em: Option<F>,
    pub w_p: Option<F>,
    pub w_m: Option<F>,

}

Data object

Computed plus and minus values are values that inherit uncertainties only from the parallax error.

Fields
outlier: bool

Is this object an outlier? (one of the discrepancies turned out to be too big)

from_reid: Option<bool>

Is this object from a Reid catalogue?

name: Option<String>

Name of the object

obj_type: Option<String>

Type of the object

source: Option<String>

Source of the data

alpha: Option<F>

Right ascension (radians)

delta: Option<F>

Declination (radians)

par: Option<F>

Parallax (mas)

par_e: Option<F>

Uncertainty in par (mas)

par_p: Option<F>

Plus value of par

par_m: Option<F>

Minus value of par

v_lsr: Option<F>

Local Standard of Rest velocity (km/s)

v_lsr_e: Option<F>

Uncertainty in v_lsr (km/s)

mu_x: Option<F>

Eastward proper motion (mas/yr)

mu_x_e: Option<F>

Uncertainty in mu_x (mas/yr)

mu_y: Option<F>

Northward proper motion (mas/yr)

mu_y_e: Option<F>

Uncertainty in mu_y (mas/yr)

r_h: Option<F>

Heliocentric distance (kpc)

r_h_ep: Option<F>

Plus uncertainty in r_h (kpc)

r_h_em: Option<F>

Minus uncertainty in r_h (kpc)

r_h_p: Option<F>

Plus value of r_h (kpc)

r_h_m: Option<F>

Minus value of r_h (kpc)

l: Option<F>

Longitude (radians)

b: Option<F>

Latitude (radians)

mu_l_cos_b: Option<F>

Proper motion in longitude

mu_b: Option<F>

Proper motion in latitude

r_g: Option<F>

Galactocentric distance (kpc)

r_g_ep: Option<F>

Plus uncertainty in r_g (kpc)

r_g_em: Option<F>

Minus uncertainty in r_g (kpc)

r_g_p: Option<F>

Plus value of r_g (kpc)

r_g_m: Option<F>

Minus value of r_g (kpc)

x: Option<F>

X coordinate (kpc)

x_ep: Option<F>

Plus uncertainty in x (kpc)

x_em: Option<F>

Minus uncertainty in x (kpc)

x_p: Option<F>

Plus value of x (kpc)

x_m: Option<F>

Minus value of x (kpc)

y: Option<F>

Y coordinate (kpc)

y_ep: Option<F>

Plus uncertainty in y (kpc)

y_em: Option<F>

Minus uncertainty in y (kpc)

y_p: Option<F>

Plus value of y (kpc)

y_m: Option<F>

Minus value of y (kpc)

z: Option<F>

Z coordinate (kpc)

z_ep: Option<F>

Plus uncertaintz in z (kpc)

z_em: Option<F>

Minus uncertaintz in z (kpc)

z_p: Option<F>

Plus value of z (kpc)

z_m: Option<F>

Minus value of z (kpc)

theta: Option<F>

Azimuthal velocity (km/s)

theta_ep: Option<F>

Plus uncertainty in theta (km/s)

theta_em: Option<F>

Minus uncertainty in theta (km/s)

theta_p: Option<F>

Plus value of theta (km/s)

theta_m: Option<F>

Minus value of theta (km/s)

theta_evel: Option<F>

Uncertainty in azimuthal velocity inherited from the velocities (km/s)

theta_evel_corrected: Option<F>

Uncertainty in azimuthal velocity inherited from the velocities (corrected) (km/s)

v_r: Option<F>

Heliocentric velocity in distance (km/s)

v_r_e: Option<F>

Uncertainty in v_r (km/s)

v_l: Option<F>

Heliocentric velocity in longitude (km/s)

v_l_ep: Option<F>

Plus uncertainty in v_l (km/s)

v_l_em: Option<F>

Minus uncertainty in v_l (km/s)

v_l_p: Option<F>

Plus value of v_l (km/s)

v_l_m: Option<F>

Minus value of v_l (km/s)

v_b: Option<F>

Heliocentric velocity in latitude (km/s)

v_b_ep: Option<F>

Plus uncertainty in v_b (km/s)

v_b_em: Option<F>

Minus uncertainty in v_b (km/s)

v_b_p: Option<F>

Plus value of v_b (km/s)

v_b_m: Option<F>

Minus value of v_b (km/s)

u: Option<F>

U coordinate (kpc)

u_ep: Option<F>

Plus uncertainty in u (kpc)

u_em: Option<F>

Minus uncertainty in u (kpc)

u_p: Option<F>

Plus value of u (kpc)

u_m: Option<F>

Minus value of u (kpc)

v: Option<F>

V coordinate (kpc)

v_ep: Option<F>

Plus uncertainty in v (km/s)

v_em: Option<F>

Minus uncertainty in v (km/s)

v_p: Option<F>

Plus value of v (km/s)

v_m: Option<F>

Minus value of v (km/s)

w: Option<F>

W coordinate (kpc)

w_ep: Option<F>

Plus uncertainty in w (km/s)

w_em: Option<F>

Minus uncertainty in w (km/s)

w_p: Option<F>

Plus value of w (km/s)

w_m: Option<F>

Minus value of w (km/s)

Implementations

impl<F> Object<F>

pub(in model) fn try_parse_alpha(
&mut self,
record: &Record<F>
) -> Result<()>where
F: Float + Debug + FromStr,
<F as FromStr>::Err: Error + Send + Sync + 'static,

Parse the right ascension string and convert the angle to radians

pub(in model) fn try_parse_delta(
&mut self,
record: &Record<F>
) -> Result<()>where
F: Float + Debug + FromStr,
<F as FromStr>::Err: Error + Send + Sync + 'static,

Parse the declination string and convert the angle to radians

impl<F> Object<F>

fn compute_x_y_z_with(&self, r_h: F) -> (F, F, F)where
F: Float + Debug,

Compute the coordinates with the specific values

pub fn compute_x_y_z(&mut self)where
F: Float + Debug,

Convert the galactic heliocentric spherical coordinates to Galactic heliocentric Cartesian coordinates

pub fn compute_u_v_w_with(&self, v_l: F, v_b: F) -> (F, F, F)where
F: Float + Debug,

Compute the velocities in the Galactic heliocentric Cartesian coordinates system (with the specific values)

pub fn compute_u_v_w_nominal(&mut self)where
F: Float + Debug + Default,

Compute the velocities in the Galactic heliocentric Cartesian coordinates system (nominal values only)

pub fn compute_u_v_w(&mut self)where
F: Float + Debug,

Compute the velocities in the Galactic heliocentric Cartesian coordinates system

impl<F> Object<F>

pub fn compute_l_b<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug,
F2: Float + Debug + Into<F>,

Convert the equatorial coordinates to spherical heliocentric Galactic coordinates

Angles must be in radians, then radians are returned.

Source: Wikipedia

pub fn compute_r_h_nominal(&mut self)where
F: Float + Debug + Default,

Compute the heliocentric distance (nominal value only)

pub fn compute_r_h(&mut self)where
F: Float + Debug,

Compute the heliocentric distance

pub fn compute_v_r<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug,
F2: Float + Debug + Into<F>,

Compute the heliocentric velocity in distance

fn compute_v_l_v_b_with<F2>(&self, r_h: F, params: &Params<F2>) -> (F, F)where
F: Float + Debug,
F2: Float + Debug + Into<F>,

Compute the velocities in longitude and latitude with the specific values

pub fn compute_v_l_v_b_nominal<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug + Default,
F2: Float + Debug + Into<F>,

Compute the velocities in longitude and latitude (nominal values only)

pub fn compute_v_l_v_b<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug,
F2: Float + Debug + Into<F>,

Compute the velocities in longitude and latitude

impl<F> Object<F>

pub fn compute_mu_l_cos_b_mu_b<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug + Default,
F2: Float + Debug + Into<F>,

Compute the proper motions in Galactic coordinates

pub fn compute_d_mu_l_cos_b_mu_b(&self, params: &Params<F>) -> (F, F)where
F: Float + Debug + Default,

Compute the dispersions of mu_l * cos(b) and mu_b

Note that only values with independent errors (from the catalog) are unpacked.

impl<F> Object<F>

fn compute_r_g_with<F2>(&self, r_h: F, params: &Params<F2>) -> Fwhere
F: Float + Debug,
F2: Float + Debug + Into<F>,

Compute the galactocentric distance with the specific values

pub fn compute_r_g_nominal<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug + Default,
F2: Float + Debug + Into<F>,

Compute the galactocentric distance (nominal value only)

pub fn compute_r_g<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug,
F2: Float + Debug + Into<F>,

Compute the galactocentric distance

impl<F> Object<F>

fn compute_theta_with<F2>(
&self,
r_h: F,
r_g: F,
u: F,
v: F,
params: &Params<F2>
) -> Fwhere
F: Float + Debug,
F2: Float + Debug + Into<F>,

Compute the azimuthal velocity with the specific values

pub fn compute_theta_nominal<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug + Default,
F2: Float + Debug + Into<F>,

Compute the azimuthal velocity (nominal value only)

pub fn compute_theta<F2>(&mut self, params: &Params<F2>)where
F: Float + Debug,
F2: Float + Debug + Into<F>,

Compute the azimuthal velocity

pub(super) fn compute_theta_evel(&mut self, params: &Params<F>)where
F: Float + Debug + Default,

Compute the uncertainty in the azimuthal velocity inherited from velocities

Note that we compute all values again, starting from the independent errors

Sources: Gromov, Nikiforov, Ossipkov (2016)

impl<F> Object<F>

pub(super) fn compute(&mut self, params: &Params<F>)where
F: Float + Debug + Default,

Perform per-object computations