Struct pmg::model::params::Params

pub struct Params<F> {

    pub r_0: F,
    pub r_0_ep: F,
    pub r_0_em: F,
    pub omega_0: F,
    pub omega_0_ep: F,
    pub omega_0_em: F,
    pub a: F,
    pub a_ep: F,
    pub a_em: F,
    pub u_sun: F,
    pub u_sun_ep: F,
    pub u_sun_em: F,
    pub v_sun: F,
    pub v_sun_ep: F,
    pub v_sun_em: F,
    pub w_sun: F,
    pub w_sun_ep: F,
    pub w_sun_em: F,
    pub sigma_r_g: F,
    pub sigma_r_g_ep: F,
    pub sigma_r_g_em: F,
    pub sigma_theta: F,
    pub sigma_theta_ep: F,
    pub sigma_theta_em: F,
    pub sigma_z: F,
    pub sigma_z_ep: F,
    pub sigma_z_em: F,
    pub theta_2: F,
    pub theta_2_ep: F,
    pub theta_2_em: F,
    pub theta_3: F,
    pub theta_3_ep: F,
    pub theta_3_em: F,
    pub theta_4: F,
    pub theta_4_ep: F,
    pub theta_4_em: F,
    pub theta_5: F,
    pub theta_5_ep: F,
    pub theta_5_em: F,
    pub theta_6: F,
    pub theta_6_ep: F,
    pub theta_6_em: F,
    pub theta_7: F,
    pub theta_7_ep: F,
    pub theta_7_em: F,
    pub theta_8: F,
    pub theta_8_ep: F,
    pub theta_8_em: F,
    pub theta_0: F,
    pub theta_0_ep: F,
    pub theta_0_em: F,
    pub theta_1: F,
    pub theta_1_ep: F,
    pub theta_1_em: F,
    pub theta_sun: F,
    pub theta_sun_ep: F,
    pub theta_sun_em: F,
    pub omega_sun: F,
    pub omega_sun_ep: F,
    pub omega_sun_em: F,
    pub alpha_ngp: F,
    pub delta_ngp: F,
    pub l_ncp: F,
    pub k: F,
    pub u_sun_standard: F,
    pub v_sun_standard: F,
    pub w_sun_standard: F,

}

Model parameters

Fields
r_0: F

Galactocentric distance to the Sun (kpc)

r_0_ep: F

Plus uncertainty in r_0

r_0_em: F

Minus uncertainty in r_0

omega_0: F

Circular velocity of the Sun at R = R_0 (km/s/kpc)

omega_0_ep: F

Plus uncertainty in omega_0

omega_0_em: F

Minus uncertainty in omega_0

a: F

Oort’s A constant (km/s/kpc)

a_ep: F

Plus uncertainty in a

a_em: F

Minus uncertainty in a

u_sun: F

Residual motion of the Sun toward GC (km/s)

u_sun_ep: F

Plus uncertainty in u_sun

u_sun_em: F

Minus uncertainty in u_sun

v_sun: F

Residual motion of the Sun toward l = 90 degrees (km/s)

v_sun_ep: F

Plus uncertainty in v_sun

v_sun_em: F

Minus uncertainty in v_sun

w_sun: F

Residual motion of the Sun toward NGP (km/s)

w_sun_ep: F

Plus uncertainty in w_sun

w_sun_em: F

Minus uncertainty in w_sun

sigma_r_g: F

Radial component of the ellipsoid of natural standard deviations (km/s)

sigma_r_g_ep: F

Plus uncertainty in sigma_r

sigma_r_g_em: F

Minus uncertainty in sigma_r

sigma_theta: F

Azimuthal component of the ellipsoid of natural standard deviations (km/s)

sigma_theta_ep: F

Plus uncertainty in sigma_theta

sigma_theta_em: F

Minus uncertainty in sigma_theta

sigma_z: F

Vertical component of the ellipsoid of natural standard deviations (km/s)

sigma_z_ep: F

Plus uncertainty in sigma_z

sigma_z_em: F

Minus uncertainty in sigma_z

theta_2: F

The second derivative of the linear rotation velocity (km/s/kpc^2)

theta_2_ep: F

Plus uncertainty in theta_2

theta_2_em: F

Minus uncertainty in theta_2

theta_3: F

The third derivative of the linear rotation velocity (km/s/kpc^3)

theta_3_ep: F

Plus uncertainty in theta_3

theta_3_em: F

Minus uncertainty in theta_3

theta_4: F

The 4th derivative of the linear rotation velocity (km/s/kpc^4)

theta_4_ep: F

Plus uncertainty in theta_4

theta_4_em: F

Minus uncertainty in theta_4

theta_5: F

The 5th derivative of the linear rotation velocity (km/s/kpc^5)

theta_5_ep: F

Plus uncertainty in theta_5

theta_5_em: F

Minus uncertainty in theta_5

theta_6: F

The 6th derivative of the linear rotation velocity (km/s/kpc^6)

theta_6_ep: F

Plus uncertainty in theta_6

theta_6_em: F

Minus uncertainty in theta_6

theta_7: F

The 7th derivative of the linear rotation velocity (km/s/kpc^7)

theta_7_ep: F

Plus uncertainty in theta_7

theta_7_em: F

Minus uncertainty in theta_7

theta_8: F

The 8th derivative of the linear rotation velocity (km/s/kpc^8)

theta_8_ep: F

Plus uncertainty in theta_8

theta_8_em: F

Minus uncertainty in theta_8

theta_0: F

The constant term of the rotation curve (km/s)

theta_0_ep: F

Plus uncertainty in theta_0

theta_0_em: F

Minus uncertainty in theta_0

theta_1: F

The first derivative of the linear rotation velocity (km/s/kpc)

theta_1_ep: F

Plus uncertainty in theta_1

theta_1_em: F

Minus uncertainty in theta_1

theta_sun: F

Linear rotation velocity of the Sun (km/s)

theta_sun_ep: F

Plus uncertainty in theta_sun

theta_sun_em: F

Minus uncertainty in theta_sun

omega_sun: F

Circular rotation velocity of the Sun (km/s/kpc)

omega_sun_ep: F

Plus uncertainty in omega_sun

omega_sun_em: F

Minus uncertainty in omega_sun

alpha_ngp: F

The right ascension of the north galactic pole (radians)

delta_ngp: F

The declination of the north galactic pole (radians)

l_ncp: F

The longitude of the north celestial pole (radians)

k: F

Linear velocities units conversion coefficient

u_sun_standard: F

Standard Solar Motion toward GC (km/s)

v_sun_standard: F

Standard Solar Motion toward l = 90 degrees (km/s)

w_sun_standard: F

Standard Solar Motion toward NGP (km/s)

Implementations

impl<F> Params<F>

pub fn update_with(&mut self, p: &[F])where
F: Float + Debug,

Update the parameters with the point in the parameter space

Note that not all fields are updated, but only those needed for fitting

pub fn to_vec(&self, n: usize, remove_sigmas: bool) -> Vec<F>where
F: Float + Debug,

Construct a point in the parameter space from the parameters

Note that not all fields are used, but only those needed for fitting

pub fn update_ep_with(&mut self, p: &[F])where
F: Float + Debug,

Update the plus uncertainties of the parameters with the values in the provided vector

Note that not all fields are updated, but only those needed for fitting

pub fn to_ep_vec(&self, n: usize) -> Vec<F>where
F: Float + Debug,

Put the plus uncertainties into a vector

pub fn update_em_with(&mut self, p: &[F])where
F: Float + Debug,

Update the minus uncertainties of the parameters with the values in the provided vector

Note that not all fields are updated, but only those needed for fitting

pub fn to_em_vec(&self, n: usize) -> Vec<F>where
F: Float + Debug,

Put the minus uncertainties into a vector

pub fn compute_with_l_stroke(index: usize, l_stroke: usize) -> bool

Should you run computations for this parameter with this L’?