Struct pmg::model::Model

pub struct Model<F> {

    pub params: Params<F>,
    pub objects: Rc<RefCell<Vec<Object<F>>>>,
    pub disable_inner: bool,
    pub lbfgs_tolerance: F,
    pub n: Option<usize>,
    pub l_stroke_3_n: Option<usize>,
    pub l_stroke_1_n: Option<usize>,
    pub best_cost: Option<F>,
    pub fit_params: Option<Params<F>>,
    pub fit_rotcurve: Option<Vec<RotationCurvePoint<F>>>,
    pub triples: Rc<RefCell<Vec<Vec<Triple<F>>>>>,
    pub covariance_matrix: Option<DMatrix<F>>,
    pub sample_description: Option<String>,
    pub output_dir: PathBuf,

}

Model of the Galaxy

Fields
params: Params<F>

Initial model parameters

objects: Rc<RefCell<Vec<Object<F>>>>

Data objects

disable_inner: bool

Disable the inner optimization?

lbfgs_tolerance: F

Tolerance of the L-BFGS algorithm

n: Option<usize>

The degree of the polynomial of the rotation curve

l_stroke_3_n: Option<usize>

Number of the objects after the L’ = 3 run

l_stroke_1_n: Option<usize>

Number of the objects after the L’ = 1 run

best_cost: Option<F>

The best value of the cost function

fit_params: Option<Params<F>>

Fit of the model (parameters)

fit_rotcurve: Option<Vec<RotationCurvePoint<F>>>

Fit of the model (rotation curve)

triples: Rc<RefCell<Vec<Vec<Triple<F>>>>>

Triples

covariance_matrix: Option<DMatrix<F>>

Covariance matrix

sample_description: Option<String>

Sample description

output_dir: PathBuf

Output directory

Implementations

impl<F> Model<F>

pub fn compute_covariance_matrix(&mut self) -> Result<()>where
F: Float + Debug + Default + Display + Sum + Sync + Send + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>> + ComplexField,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F> + FiniteDiff<F>,

Compute the covariance matrix

impl<F> Model<F>where
F: Float + Debug + Default + Display + Sync + Send + Sum + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F> + FiniteDiff<F>,

pub fn try_fit_errors(
&mut self,
errors_log_writer: &Rc<RefCell<BufWriter<File>>>,
l_stroke: usize
) -> Result<()>

Try to define the confidence intervals

fn try_fit_errors_pair<FN>(
&self,
l_stroke: usize,
index: usize,
param: F,
compute_param: FN,
errors_log_writer: &Rc<RefCell<BufWriter<File>>>,
triples: &Rc<RefCell<Vec<Vec<Triple<F>>>>>
) -> Result<(F, F)>where
FN: Fn(F, &[F]) -> F,

Try to define a confidence interval of one parameter

impl<F> Model<F>

pub fn analyze_inner_profiles(&self) -> Result<()>where
F: Float + Debug + Default + Display + Sync + Send,

Compute the profiles of the inner target function. Also, find those objects that have profiles with multiple local minima and output their coordinates

impl<F> Model<F>

pub fn find_outliers(
&mut self,
l_stroke: usize
) -> Result<(OneDimensionalOutliers<F>, MultiDimensionalOutliers<F>)>where
F: Float + Debug + Default + Sync + Send + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F>,

Find outliers

pub fn find_outliers_1d(
&mut self,
n_nonblacklisted: usize,
l_stroke: usize
) -> Result<OneDimensionalOutliers<F>>where
F: Float + Debug + Default + Sync + Send + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F>,

Find outliers using the one-dimensional algorithm

pub fn find_outliers_3d(
&mut self,
n_nonblacklisted: usize,
l_stroke: usize
) -> Result<MultiDimensionalOutliers<F>>where
F: Float + Debug + Default + Sync + Send + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F>,

Find outliers using the three-dimensional algorithm

pub fn find_outliers_4d(
&mut self,
n_nonblacklisted: usize,
l_stroke: usize
) -> Result<MultiDimensionalOutliers<F>>where
F: Float + Debug + Default + Sync + Send + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F>,

Find outliers using the four-dimensional algorithm

impl<F> Model<F>

pub fn write_parallaxes(&self) -> Result<()>where
F: Float + Debug + Display,

Write the parallaxes (original and reduced ones)

pub fn write_delta_varpi(
&self,
plain_writer: &mut BufWriter<File>,
dat_writer: &mut BufWriter<File>
) -> Result<()>where
F: Float + Debug + Display,

Compute the systematical error in parallaxes

impl<F> Model<F>

pub fn try_fit_params(
&mut self,
n: usize,
sample_iteration: usize,
l_stroke: usize,
fit_log_writer: &Rc<RefCell<BufWriter<File>>>
) -> Result<()>where
F: Float + Debug + Default + Display + Sync + Send + Sum + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F> + FiniteDiff<F>,

Try to fit the model of the Galaxy to the data

impl<F> Model<F>

pub fn try_compute_conditional_profiles(&mut self, l_stroke: usize) -> Result<()>where
F: Float + Debug + Default + Display + Sync + Send + Sum + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F> + FiniteDiff<F>,

Try to compute the conditional profiles

fn try_compute_conditional_profile<FN>(
&mut self,
l_stroke: usize,
index: usize,
compute_param: FN,
fit_param: F,
fit_param_ep: F,
fit_param_em: F,
triples: &Rc<RefCell<Vec<Vec<Triple<F>>>>>
) -> Result<Vec<ProfilePoint<F>>>where
F: Float + Debug + Default + Display + Sync + Send + Sum + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F> + FiniteDiff<F>,
FN: Fn(F, &[F]) -> F,

Try to compute a conditional profile

pub fn try_compute_frozen_profiles(&mut self, l_stroke: usize) -> Result<()>where
F: Float + Debug + Default + Display + Sync + Send + Sum + ArgminFloat + ArgminL2Norm<F> + ArgminSub<F, F> + ArgminAdd<F, F> + ArgminDot<F, F> + ArgminMul<F, F> + ArgminZeroLike + ArgminMul<Vec<F>, Vec<F>>,
Vec<F>: ArgminSub<Vec<F>, Vec<F>> + ArgminSub<F, Vec<F>> + ArgminAdd<Vec<F>, Vec<F>> + ArgminAdd<F, Vec<F>> + ArgminMul<F, Vec<F>> + ArgminMul<Vec<F>, Vec<F>> + ArgminL1Norm<F> + ArgminSignum + ArgminMinMax + ArgminDot<Vec<F>, F> + ArgminL2Norm<F> + FiniteDiff<F>,

Try to compute the frozen profiles

pub fn serialize_to_profile(
&self,
profile_type: &ProfileType,
profile: &Vec<ProfilePoint<F>>,
param_name: &str
) -> Result<()>where
F: Float + Debug + Display + Serialize,

Serialize the profiles

impl<F> Model<F>

pub fn compute_fit_rotcurve(&mut self)where
F: Float + Debug,

Compute the rotation curve based on the fitted parameters

pub fn serialize_to_fit_rotcurve(&self) -> Result<()>where
F: Float + Debug + Display + Serialize,

Serialize the fitted rotation curve

impl<F> Model<F>

pub fn count_non_outliers(&self) -> usize

Count the number of non-outliers

pub fn get_outliers_mask(&self) -> Vec<bool>

Get the mask with the outliers

pub fn apply_outliers_mask(&self, mask: &[bool])

Apply the outliers mask

pub fn try_load_data_from(&mut self, path: &Path) -> Result<()>where
F: Float + Default + Debug + DeserializeOwned + FromStr,
<F as FromStr>::Err: Error + Send + Sync + 'static,

Try to load data from the path

pub fn serialize_to_objects(&self, name: &str, params: &Params<F>) -> Result<()>where
F: Float + Debug + Display + Serialize,

Serialize the per-object data

impl<F> Model<F>

pub(super) fn serialize_to_params(&self) -> Result<()>where
F: Float + Debug + Display + Serialize,

Serialize the initial parameters

pub fn serialize_to_fit_params(&self) -> Result<()>where
F: Float + Debug + Display + Serialize,

Serialize the fitted parameters

pub fn write_fit_params_header_to_plain(
&self,
plain_writer: &mut BufWriter<File>
) -> Result<()>where
F: Display,

Write the header to the plain file

pub fn write_fit_params_footer_to_plain(
&self,
plain_writer: &mut BufWriter<File>
) -> Result<()>where
F: Display,

Write the footer to the plain file

pub fn write_fit_params_to_plain(
&self,
plain_writer: &mut BufWriter<File>,
n: usize
) -> Result<()>where
F: Float + Debug + Display,

Represent the fitted parameters in a plain view

impl<F> Model<F>

pub fn try_read_sample_description_from(&mut self, path: &Path) -> Result<()>

Read the sample description from the input file

pub fn format_sample_description(&self) -> String

Get the sample description as a formatted string

impl<F> Model<F>

pub fn compute_objects(&mut self)where
F: Float + Debug + Default,

Perform per-object computations

pub fn post_fit(&mut self)where
F: Float + Debug + Default,

Run the post-fit computations

pub fn write_objects_data(&self) -> Result<()>where
F: Float + Debug + Display + Serialize,

Write the objects data

pub fn write_fit_data(&self) -> Result<()>where
F: Float + Debug + Display + Serialize,

Write the fit data

pub fn try_from(args: &Args, output_dir: PathBuf) -> Result<Self>where
F: Float + Debug + Default + DeserializeOwned + FromStr,
<F as FromStr>::Err: Error + Send + Sync + 'static,

Create a model from the arguments with a specific degree of the polynomial of the rotation curve