Module pmg::utils

Utilities

Re-exports

pub use compute_e_theta::compute_e_theta;

pub use compute_r_g::compute_r_g_1;

pub use compute_r_g::compute_r_g_2;

pub use compute_theta_r_g::compute_theta_r_g;

pub use dms2rad::dms2rad;

pub use hms2rad::hms2rad;

pub use str2vec::str2vec;

pub use to_cartesian::to_cartesian;

pub use to_spherical::to_spherical;

Modules

compute_e_theta 🔒

Compute the uncertainty of the azimuthal velocity

compute_r_g 🔒

Compute the distance in the Galactocentric coordinate system associated with the object

compute_theta_r_g 🔒

Compute the azimuthal velocity and Galactocentric distance

dms2rad 🔒

Convert a degrees-minutes-seconds angle to radians

hms2rad 🔒

Convert an hours-minutes-seconds angle to radians

str2vec 🔒

Parse a string for a vector of floats

to_cartesian 🔒

Convert heliocentric Galactic coordinates from the spherical to the Cartesian coordinate system

to_spherical 🔒

Convert equatorial coordinates to spherical Galactic coordinates